planeUpgrade trigger and changed name of Schema
planeUpgrade trigger creates properly but gets a compiler warning.
warning states a missing right parenthesis but there isn't one missing.
It's something to do with the IF statement in there if somebody wants to
take a look.

Can't test if it works regardless of the warning until data is made